+++ src/test/java/com/hhplu/hhplusconcert/app/application/unittest/QueueServiceTest.java
@@ -1,4 +1,4 @@
-package com.hhplu.hhplusconcert.app.application.token;
+package com.hhplu.hhplusconcert.app.application.unittest;
 
 import com.hhplu.hhplusconcert.app.application.waitingqueue.service.WaitingQueueService;
 import com.hhplu.hhplusconcert.app.domain.watingqueue.WaitingQueueStatus;
