+++ src/test/java/com/hhplu/hhplusconcert/app/application/unittest/ReservationServiceTest.java
@@ -1,5 +1,6 @@
-package com.hhplu.hhplusconcert.app.application.reservation.service;
+package com.hhplu.hhplusconcert.app.application.unittest;
 
+import com.hhplu.hhplusconcert.app.application.reservation.service.ReservationService;
 import com.hhplu.hhplusconcert.app.domain.reservation.ReservationStatus;
 import com.hhplu.hhplusconcert.app.domain.reservation.entity.Reservation;
 import com.hhplu.hhplusconcert.app.domain.reservation.repository.ReservationRepository;
