+++ src/test/java/com/hhplu/hhplusconcert/app/application/unittest/SeatServiceTest.java
@@ -1,4 +1,4 @@
-package com.hhplu.hhplusconcert.app.application.token;
+package com.hhplu.hhplusconcert.app.application.unittest;
 
 import com.hhplu.hhplusconcert.app.application.concert.service.SeatService;
 import com.hhplu.hhplusconcert.app.domain.concert.SeatStatus;
